🩹 Corrige la base de données
