+++ src/public/js/agenda/agendaMenu.js
@@ -1,11 +1,14 @@
+// Fonction pour afficher/masquer le menu
 function toggleMenu() {
   const dropdownMenu = document.getElementById("dropdown-menu");
   const isVisible = dropdownMenu.style.display === "block";
-
   dropdownMenu.style.display = isVisible ? "none" : "block";
 }
 
-// Ferme le menu lorsque l'utilisateur clique en dehors de celui-ci
+// Attacher l'événement au bouton menu
+document.getElementById("menu-button").addEventListener("click", toggleMenu);
+
+// Fermer le menu lorsque l'utilisateur clique en dehors
 document.addEventListener("click", function (event) {
   const dropdownMenu = document.getElementById("dropdown-menu");
   const menuButton = document.getElementById("menu-button");
