--- src/public/js/logout.js
@@ -1,7 +1,0 @@
-/**
- * Envoie une requête POST pour déconnecter l'utilisateur
- */
-// eslint-disable-next-line no-unused-vars
-function sendLogout() {
-  document.getElementById("logoutForm").submit();
-}
