+++ src/views/account.ejs
@@ -87,7 +87,7 @@
 
 
     <% if (locals.user) { %>
-        <script src="/js/accountEdition.js"></script>
+        <script type="module" src="/js/account/accountEdition.js"></script>
     <% } %>
 
     <%- include('partials/footer') %>
