+++ src/views/calendar.ejs
@@ -3,7 +3,7 @@
 <script src='https://cdn.jsdelivr.net/npm/fullcalendar@6.1.15/index.global.min.js'></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.29.1/moment.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.29.1/locale/fr.min.js"></script>
-<script src="/js/calendar.js"></script>
+<script type="module" src="/js/agenda/calendar.js"></script>
 
 <!-- Modale pour afficher les informations de l'événement -->
 <div id="eventModal" class="modal">
