+++ src/views/login.ejs
@@ -37,7 +37,7 @@
         </div>
     </div>
 
-    <script src="/js/login.js"></script>
+    <script type="module" src="/js/account/login.js"></script>
 
     <%- include('partials/footer') %>
 </body>
