+++ src/views/signin.ejs
@@ -49,7 +49,7 @@
     </div>
 
 
-    <script type="module" src="/js/register.js"></script>
+    <script type="module" src="/js/account/register.js"></script>
 
     <%- include('partials/footer') %>
 </body>
